Rewrite cache tests.
